requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
matplotlib==3.8.1
seaborn==0.13.0
plotly==5.17.0
pandas==2.1.3
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
